pyproject.toml
[tool.poetry]
name = "luna_backend"
version = "0.1.0"
description = "The Discord-bot for Dusked Ocean server."
authors = ["Twylix <zemskov271@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.6"
gunicorn = "^20.1.0"
requests = "^2.28.1"
pydantic = "^1.9.1"
psycopg2 = "^2.9.3"
drf-yasg = "^1.21.3"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
isort = "^5.10.1"
wemake-python-styleguide = "^0.16.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output = 3
skip_glob=[".venv/*", "docker/*", "shared/*", "ssl/*"]
include_trailing_comma = true

[tool.black]
skip-string-normalization = true
line-length = 90
exclude = '(\.git|\.venv|\.vscode|\docker|\shared|\ssl)'
